middle stagge, unfinished
